Move code into respective modules
